working eurostat scraper

+++ src/scrapers/eurostat/eurostat_restsdmx.R
@@ -15,39 +15,52 @@ dir.create(outdir, showWarnings=F, recursive=T)
 source(file.path(srcdir, 'functions.R'))
 
 
-# get Eurostat catalogue
-ec_url <- 'https://ec.europa.eu/eurostat/api/dissemination/sdmx/2.1/dataflow/ESTAT/all?detail=allstubs'
-eurostat_catalogue <- as.data.frame(readSDMX(ec_url))
-
-# data-of-interest
-endpoints <- c('MEDPS12')
-eurostat_catalogue$mapineq <- as.numeric(eurostat_catalogue$id %in% endpoints)
-
-# save catalogue
-write.csv(eurostat_catalogue, file.path(outdir, 'eurostat_catalogue.csv'), row.names=F)
-
-
-# get xml metadata
-mdat <- list()
-for(endpoint in endpoints){
+if(file.exists(file.path(indir, 'eurostat_catalogue.csv'))){
   
-  # get metadata
-  mdat[[endpoint]] <- get_xml(endpoint = endpoint,
-                              datdir = file.path(outdir, 'data'),
-                              overwrite = FALSE)
+  # load eurostat catalogue
+  eurostat_catalogue <- read.csv(file.path(indir, 'eurostat_catalogue.csv'))
+  
+  # identify data-of-interest from prepared catalogue
+  endpoints <- eurostat_catalogue$id[eurostat_catalogue$mapineq==1 & !is.na(eurostat_catalogue$mapineq)]
+  
+} else {
+  
+  # get eurostat catalogue
+  ec_url <- 'https://ec.europa.eu/eurostat/api/dissemination/sdmx/2.1/dataflow/ESTAT/all?detail=allstubs'
+  eurostat_catalogue <- as.data.frame(readSDMX(ec_url))
+  
+  # identify data-of-interest in this list
+  endpoints <- c('')
+  eurostat_catalogue$mapineq <- as.numeric(eurostat_catalogue$id %in% endpoints)
+  
+  # save catalogue
+  write.csv(eurostat_catalogue, file.path(outdir, 'eurostat_catalogue.csv'), row.names=F)
 }
 
-# get indicator lists
-indicators <- list()
-for(endpoint in endpoints){
-  
-  # get metadata
-  indicators[[endpoint]] <- get_indicators(endpoint = endpoint,
-                                           datdir = file.path(outdir, 'data'),
-                                           overwrite = FALSE)
-}
-codebook <- build_codebook(datdir = file.path(outdir, 'data'))
-write.csv(codebook, file = file.path(outdir, 'codebook.csv'), row.names = FALSE)
+
+# # get xml metadata
+# mdat <- list()
+# for(endpoint in endpoints){
+#   
+#   # get metadata
+#   mdat[[endpoint]] <- get_xml(endpoint = endpoint,
+#                               datdir = file.path(outdir, 'data'),
+#                               overwrite = FALSE)
+# }
+
+
+# # get indicator lists
+# indicators <- list()
+# for(endpoint in endpoints){
+#   
+#   # get metadata
+#   indicators[[endpoint]] <- get_indicators(endpoint = endpoint,
+#                                            datdir = file.path(outdir, 'data'),
+#                                            overwrite = FALSE)
+# }
+# codebook <- build_codebook(datdir = file.path(outdir, 'data'))
+# write.csv(codebook, file = file.path(outdir, 'codebook.csv'), row.names = FALSE)
+
 
 # get data
 dat <- list()
